FAIR-2507 bug修改
