+++ scripts/.conf
@@ -90,7 +90,7 @@
 			]
 		},
 		{
-			"path":"../../gaia_wallet_webserver_rel/plugin/wallet/htdocs/.res/wallet",
+			"path":"../../gaia_wallet_webserver_vm/plugin/wallet/htdocs/.res/wallet",
 			"pre":"rel",
 			"ignoresuf":[".depend",".r.ts"],
 			"ignoredir":["src/boot/"],
@@ -258,7 +258,7 @@
 			]
 		},
 		{
-			"path":"../../gaia_wallet_webserver_vm/plugin/wallet/htdocs/.res/wallet",
+			"path":"../../gaia_wallet_webserver_vm/plugin/wallet/htdocs/.res/wallet_browser",
 			"pre":"browser",
 			"ignoresuf":[".depend",".r.ts"],
 			"ignoredir":["src/app/core/","src/app/remote/","src/app/store/","src/boot/"],
