+++ src/app/boot/next.js
@@ -254,9 +254,9 @@ winit.initNext = function () {
 			console.log("stage webview goReady");
 			WebViewManager.getReady("firstStage");   // 通知一阶段准备完毕
 			// 继续加载首页所需
+			loadWalletFirstPageSource();  //钱包
 			loadChatSource();  // 聊天
 			loadEarnSource();  // 活动
-			loadWalletFirstPageSource();  //钱包
 			loadImages(); // 预加载图片
 			if(!pi_update.inApp){
 				vmLoaded();
@@ -310,7 +310,6 @@ winit.initNext = function () {
 			"app/view/base/",
 			"app/components1/",
 			"app/res/css/",
-			"app/res/js/",
 			"app/view/play/home/",
 			"app/view/wallet/home/",
 			"earn/client/app/res/css/",
@@ -326,7 +325,6 @@ winit.initNext = function () {
 			tab.timeout = 90000;
 			tab.release();
 			fpFlags.walletReady = true;
-			
 
 			enterApp();
 			
@@ -346,8 +344,7 @@ winit.initNext = function () {
 			"chat/client/app/view/contactList/contactItem.tpl",
 			"chat/client/app/view/contactList/contactItem.js",
 			"chat/client/app/view/contactList/contactItem.wcss",
-			"chat/client/app/widget/imgShow/",
-			"chat/client/app/widget/topBar/"
+			"chat/client/app/widget/imgShow/"
 		]; 
 		util.loadDir(sourceList, flags, fm, suffixCfg, function (fileMap) {
 			console.timeEnd("fp loadChatSource");
@@ -358,7 +355,6 @@ winit.initNext = function () {
 			tab.release();
 			fpFlags.chatReady = true;
 			enterApp();
-			
 		}, function (r) {
 			alert("加载目录失败, " + r.error + ":" + r.reason);
 		},dirProcess.handler);
@@ -386,7 +382,6 @@ winit.initNext = function () {
 			console.log("load loadEarnSource-----------------");
 			fpFlags.earnReady = true;
 			enterApp();
-			
 		}, function (r) {
 			alert("加载目录失败, " + r.error + ":" + r.reason);
 		}, dirProcess.handler);
