+++ src/app/view/mine/account/home.ts
@@ -7,8 +7,7 @@ import { Forelet } from '../../../../pi/widget/forelet';
 import { Widget } from '../../../../pi/widget/widget';
 import { getStoreData } from '../../../middleLayer/wrap';
 import { uploadFile } from '../../../net/pull';
-import { changeWalletName, walletNameAvailable } from '../../../utils/account';
-import { getUserInfo, imgResize, popNewMessage, popPswBox, rippleShow } from '../../../utils/tools';
+import { changeWalletName, getUserInfo, imgResize, popNewMessage, popPswBox, rippleShow, walletNameAvailable } from '../../../utils/tools';
 import { registerStoreData } from '../../../viewLogic/common';
 import { exportMnemonic } from '../../../viewLogic/localWallet';
 import { selectImage } from '../../../viewLogic/native';
