+++ src/app/view/mine/setting/changePsw.ts
@@ -6,8 +6,7 @@ import { getLang } from '../../../../pi/util/lang';
 import { Forelet } from '../../../../pi/widget/forelet';
 import { Widget } from '../../../../pi/widget/widget';
 import { callPasswordChange, callVerifyIdentidy } from '../../../middleLayer/wrap';
-import { pswEqualed } from '../../../utils/account';
-import { popNewLoading, popNewMessage } from '../../../utils/tools';
+import { popNewLoading, popNewMessage, pswEqualed } from '../../../utils/tools';
 // ================================ 导出
 // tslint:disable-next-line:no-reserved-keywords
 declare var module: any;
