+++ src/app/view/mine/setting/settingPsw.ts
@@ -4,10 +4,9 @@
 import { popNew } from '../../../../pi/ui/root';
 import { getLang } from '../../../../pi/util/lang';
 import { Widget } from '../../../../pi/widget/widget';
+import { defaultPassword } from '../../../config';
 import { callPasswordChange, callVerifyIdentidy } from '../../../middleLayer/wrap';
-import { pswEqualed } from '../../../utils/account';
-import { defaultPassword } from '../../../utils/constants';
-import { getUserInfo, popNewLoading, popNewMessage } from '../../../utils/tools';
+import { getUserInfo, popNewLoading, popNewMessage, pswEqualed } from '../../../utils/tools';
 
 export class SettingPsw extends Widget {
     public props: any;
