+++ src/app/view/wallet/transaction/transactionDetails.ts
@@ -6,10 +6,10 @@ import { popNew } from '../../../../pi/ui/root';
 import { getLang } from '../../../../pi/util/lang';
 import { Forelet } from '../../../../pi/widget/forelet';
 import { Widget } from '../../../../pi/widget/widget';
+import { blockchainUrl, etherscanUrl } from '../../../config';
 import { getStoreData } from '../../../middleLayer/wrap';
 import { CurrencyRecord, TxType } from '../../../publicLib/interface';
 import { timestampFormat } from '../../../publicLib/tools';
-import { blockchainUrl, etherscanUrl } from '../../../utils/constants';
 import { canResend, copyToClipboard, fetchLocalTxByHash1, parseAccount, parseStatusShow, popNewMessage } from '../../../utils/tools';
 import { registerStoreData } from '../../../viewLogic/common';
 import { makeScreenShot, openNewActivity } from '../../../viewLogic/native';
